Changes vsId to INT not BIGINT as it is not a primary key in the db

+++ DB/ACVP_DB/acvp/Stored Procedures/ErrorPut.sql
@@ -4,7 +4,7 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE PROCEDURE [acvp].[ErrorPut]
-    @VsID BIGINT,
+    @VsID INT,
     @Error NVARCHAR(MAX)
 AS
     UPDATE [acvp].[VectorSetJson]
